Rename ReactiveBaseComponent
