+++ src/app/reactive-templates/reactive-components/components/nested-template/nested-template.component.html
@@ -1,1 +1,5 @@
-<oab-reactive-template [templateName]="value()" [namespace]="name()"></oab-reactive-template>
+<oab-reactive-template
+  [templateName]="value()"
+  [namespace]="name()"
+  [attr.data-test]="name()"
+></oab-reactive-template>
